+++ src/libtrx/include/libtrx/game/inventory_ring/control.h
@@ -4,9 +4,10 @@
 #include "../objects/types.h"
 #include "./types.h"
 
-INV_RING *InvRing_Open(INVENTORY_MODE mode);
-void InvRing_Close(INV_RING *ring);
+extern INV_RING *InvRing_Open(INVENTORY_MODE mode);
+extern void InvRing_Close(INV_RING *ring);
 
-GF_COMMAND InvRing_Control(INV_RING *ring, int32_t num_frames);
+extern GF_COMMAND InvRing_Control(INV_RING *ring, int32_t num_frames);
 
+void InvRing_AdjustMusicVolume(const INV_RING *ring);
 void InvRing_SetRequestedObjectID(GAME_OBJECT_ID obj_id);